Delete unused oss_case, move ImportErrorHandler to mars.utils and reformat code.
